chatops/chatops_analyst/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.35.2
torch>=2.0.0
accelerate==0.24.1
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
pandas==2.1.4
numpy==1.24.3
pydantic==2.5.0
